backend/requirements.txt
# Core Dependencies
python-dateutil>=2.8.2
pytz>=2023.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0
lightgbm>=3.4.0

# Deep Learning & LLM
torch>=2.0.0
transformers>=4.30.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.6.1
motor>=3.3.2

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.65.0
google-generativeai>=0.3.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Monitoring
prometheus-client>=0.17.0
